chore: move selection logic to canvas context
